[service:tracing] Create TracingSingleton
